chore: more refactors
