+++ server/models/base.ts
@@ -1,13 +1,6 @@
 import { log } from 'loglevel'
 import { CaptureFeature } from 'models/captureFeature'
-import knex from 'services/knex'
-
-export enum TableNames {
-  CAPTURE_FEATURE = 'capture_feature',
-  RAW_CAPTURE_FEATRURE = 'raw_capture_feature',
-  RAW_CAPTURE_CLUSTER = 'raw_capture_cluster',
-  REGION_ASSIGNMENT = 'region_assignment',
-}
+import knex, { TableNames } from 'services/knex'
 
 export async function batchUpdate(
   ids: string[],
